requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
redis==5.0.1
celery==5.3.4